Use golang.org/x/sys/unix to work around syscall differences
