+++ constants_linux.go
@@ -1,0 +1,8 @@
+package termenv
+
+import "golang.org/x/sys/unix"
+
+const (
+	tcgetattr = unix.TCGETA
+	tcsetattr = unix.TCSETA
+)
