+++ go.mod
@@ -3,8 +3,8 @@ module github.com/muesli/termenv
 go 1.13
 
 require (
-	github.com/google/goterm v0.0.0-20190703233501-fc88cf888a3f
 	github.com/lucasb-eyer/go-colorful v1.0.3
 	github.com/mattn/go-isatty v0.0.12
 	github.com/mattn/go-runewidth v0.0.9
+	golang.org/x/sys v0.0.0-20200116001909-b77594299b42
 )
